+++ package.json
@@ -4,6 +4,7 @@
     "scripts": {
         "dev": "vite",
         "build": "vite build",
+        "watch": "vite build --watch",
         "start": "vite"
     },
     "devDependencies": {
